ajout source postman, tu delete

+++ src/main/java/com/xoff/pokepetal/service/PokemonServiceImpl.java
@@ -1,5 +1,6 @@
 package com.xoff.pokepetal.service;
 
+import com.xoff.pokepetal.dto.PageMapper;
 import com.xoff.pokepetal.dto.PagePokemonDto;
 import com.xoff.pokepetal.dto.PokemonDto;
 import com.xoff.pokepetal.dto.PokemonMapper;
@@ -8,7 +9,6 @@ import com.xoff.pokepetal.model.Pokemon;
 import com.xoff.pokepetal.repository.PokemonRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.http.HttpStatus;
@@ -33,30 +33,32 @@ public class PokemonServiceImpl implements PokemonService {
     }
 
     @Override
-    public PokemonDto findPokemonById(Long id) throws BusinessException {
+    public PokemonDto findPokemonById(Long id) {
         Optional<Pokemon> pokemonFound = pokemonRepository.findById(id);
         if (Boolean.FALSE.equals(pokemonFound.isPresent())) {
-           return null;
-        }else{
-        return PokemonMapper.INSTANCE.pokemon2PokemonDto(pokemonFound.get());}
-    }
-
-    @Transactional
-    @Override
-    public void deletePokemon(Long id) throws BusinessException {
-        try {
-            pokemonRepository.deleteById(id);
-        } catch (EmptyResultDataAccessException ex) {
-            log.error(String.format("Aucun Pokemon n'existe avec l'identifiant: " + id, ex));
-            throw new BusinessException("DeleteUserError", "Erreur de suppression du Pokemon avec l'identifiant: " + id, HttpStatus.NOT_FOUND);
-        } catch (Exception ex) {
-            throw new BusinessException("DeleteUserError", "Erreur de suppression du Pokemonavec l'identifiant: " + id, HttpStatus.INTERNAL_SERVER_ERROR);
+            return null;
+        } else {
+            return PokemonMapper.INSTANCE.pokemon2PokemonDto(pokemonFound.get());
         }
     }
 
     @Transactional
     @Override
-    public PokemonDto saveOrUpdate(PokemonDto pokemonDto) throws BusinessException {
+    public boolean deletePokemon(Long id) {
+
+            PokemonDto pokemonDto=findPokemonById(id);
+            if (pokemonDto!=null) {
+                pokemonRepository.deleteById(id);
+                return true;
+            }else{
+                return false;
+            }
+
+    }
+
+    @Transactional
+    @Override
+    public PokemonDto saveOrUpdate(PokemonDto pokemonDto) {
         try {
 
             Pokemon pokemonToSave = PokemonMapper.INSTANCE.pokemonDto2Pokemon(pokemonDto);
@@ -71,10 +73,7 @@ public class PokemonServiceImpl implements PokemonService {
 
     public PagePokemonDto findAll(Pageable paging) {
         Page<Pokemon> pagePokemon = pokemonRepository.findAll(paging);
-        PagePokemonDto pagePokemonDto = new PagePokemonDto();
-        pagePokemonDto.setTotalPages(pagePokemon.getTotalPages());
-        pagePokemonDto.setNumber(pagePokemon.getNumber());
-        pagePokemonDto.setTotalElements(pagePokemon.getTotalElements());
+        PagePokemonDto pagePokemonDto = PageMapper.INSTANCE.page2PageDto(pagePokemon);
         pagePokemonDto.setListPokemonsDto(PokemonMapper.INSTANCE.mapListEntity2Dto(pagePokemon.getContent()));
         return pagePokemonDto;
     }
